refactor: move Purpose and CountdownCallback to separate modules
Refactor `Purpose` and `CountdownCallback` classes into separate modules.
This change improves code organization by moving `Purpose` to `purpose.py` and `CountdownCallback` to `callback.py`.
