+++ src/app.py
@@ -47,16 +47,8 @@ from src import time_utils
 from src import constants
 from src import platforms
 from src.dfont import DFont
-
-
-@dataclass
-class Purpose:
-    text: str
-    timestamp: float = dataclasses.field(default_factory=time)
-
-    def add_to_history(self, history_file: Path):
-        with history_file.open("a") as f:
-            f.write(json.dumps(self.__dict__) + "\n")
+from src.purpose import Purpose
+from src.callback import CountdownCallback
 
 
 class Scene(Enum):
@@ -113,29 +105,6 @@ class Scene(Enum):
             )
 
         return rects
-
-
-class CountdownCallback:
-    """Call a command once the timer goes below a specific time."""
-
-    def __init__(self, time_and_command: str) -> None:
-        time, _, command = time_and_command.partition(":")
-        self.command = command
-        if isinstance(time, str):
-            self.time = time_utils.human_duration(time)
-        else:
-            self.time = time
-        self.executed = False
-
-    def update(self, current_time: float):
-        """Call the command if needed. Current time is the number of seconds on screen."""
-        if current_time >= self.time:
-            self.executed = False
-        elif not self.executed:
-            self.executed = True
-            # Asynchronously run the command.
-            print(f"Running: {self.command}")
-            subprocess.Popen(self.command, shell=True)
 
 
 app = typer.Typer(add_completion=False)
